Working

+++ gateway/storage_gw.go
@@ -1,6 +1,8 @@
 package gateway
 
 import (
+	"errors"
+	"fmt"
 	"io/ioutil"
 	"os"
 )
@@ -21,9 +23,12 @@ func NewStorageGateway(storagePath string) *StorageGateway {
 
 func (gw *StorageGateway) SaveDocument(path string, bytes []byte) error {
 
-	err := ioutil.WriteFile(path, bytes, 0)
-	if err != nil {
-		return err
+	if _, err := os.Stat(path); errors.Is(err, os.ErrNotExist) {
+		err := ioutil.WriteFile(path, bytes, 0)
+		if err != nil {
+			fmt.Println(err)
+			return err
+		}
 	}
 	return nil
 }
